Rename get-vue-apps and add getVueComponents.
