--- Assets/Scripts/helpers/get-vue-apps.js
@@ -1,6 +1,0 @@
-const path = require('path');
-const fs = require('fs');
-
-module.exports = (rootPath) => fs
-    .readdirSync(rootPath)
-    .filter((file) => fs.statSync(path.join(rootPath, file)).isDirectory());
